+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 Features:
 - Added the `RouteProvider.tag` constant to reduce magic symbols.
+- Added a ServiceProvider extension for registering the explorator services.
 
 ## 1.0.0